INTERLOK-3162 Add a new jsonlib-transformation-driver
- SimpleTransformationDriver is now the default; this won't break
existing config since it was always set in the constructor anyway
- DefaultTransformationDriver has been deprecated in favour of
JsonlibTransformationDriver
- JsonArray / JsonObject now extends JsonlibTransformationDriver
- Fix tests for hierarchy change.
