+++ src/test/java/com/adaptris/core/transform/json/JsonObjectTransformationDriverTest.java
@@ -4,13 +4,9 @@ import static org.junit.Assert.fail;
 import org.junit.Test;
 import com.adaptris.core.ServiceException;
 
-public class JsonObjectTransformationDriverTest extends DefaultTransformDriverTest {
+public class JsonObjectTransformationDriverTest extends JsonlibTransformDriverTest {
 
 
-  @Override
-  public boolean isAnnotatedForJunit4() {
-    return true;
-  }
   @Override
   @Test
   public void testJsonArrayToXml() throws Exception {
